backend/requirements.txt
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
eland==8.12.1
elastic-transport==8.12.0
elasticsearch==8.6.2
exif==1.5.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.0.1
fonttools==4.49.0
fsspec==2024.2.0
gunicorn==23.0.0
huggingface-hub==0.21.3
idna==3.6
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jwt==1.3.1
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==1.5.3
Pillow==9.3.0
plum-py==0.8.7
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.9.0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
SQLAlchemy==2.0.32
sympy==1.13.2
threadpoolctl==3.3.0
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.64.1
transformers==4.40.1
typing_extensions==4.10.0
ultralytics==8.2.103
ultralytics-thop==2.0.8
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.0.1
zipp==3.17.0
